crewai/pyproject.toml
[project]
name = "ai-oda-crewai"
version = "0.1.0"
description = "CrewAI content strategy agents for AI-ODA social media automation"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.14.0",
    "redis>=5.0.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
crewai-bridge = "src.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
